requirements.txt
smartsheet-python-sdk==3.0.0
snowflake-connector-python>=3.0.0,<4.0.0
requests>=2.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
json5>=0.9.0
openpyxl>=3.0.0
cryptography>=40.0.0 